Makes Java Timer timing more accurate
Java was using integer division for calulating Timer values, which
was causing Timer values to be concatinated, and only accurate
to the nearest millisecond. This changes it to use double division,
which matches C++ and makes timing much more accurate

+++ wpilibj/src/athena/java/edu/wpi/first/wpilibj/internal/HardwareTimer.java
@@ -58,7 +58,7 @@ public class HardwareTimer implements Timer.StaticInterface {
   }
 
   class TimerImpl implements Timer.Interface {
-    private long m_startTime;
+    private double m_startTime;
     private double m_accumulatedTime;
     private boolean m_running;
 
@@ -70,8 +70,8 @@ public class HardwareTimer implements Timer.StaticInterface {
       reset();
     }
 
-    private long getMsClock() {
-      return Utility.getFPGATime() / 1000;
+    private double getMsClock() {
+      return Utility.getFPGATime() / 1000.0;
     }
 
     /**
